# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A CLI tool and API server for managing LLM resources, containers, and RAG pipelines.",
        "title": "llm-manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "PLACEHOLDER_HOST",
    "basePath": "/api",
    "paths": {
        "/models": {
            "get": {
                "description": "Returns all LLM models. When OData query parameters are present ($filter, $search, $sort, $page, $limit, $fields), returns a wrapped OData response with pagination metadata. Without OData params, returns a flat JSON array for backward compatibility.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "List all models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter expression (e.g. 'type eq llm')",
                        "name": "$filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Free-text search on name and slug",
                        "name": "$search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (e.g. 'name asc')",
                        "name": "$sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "$page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "$limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of fields to include",
                        "name": "$fields",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OData-wrapped response when OData params present",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ODataListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new LLM model from a JSON body. The model slug must be unique.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Create a new model",
                "parameters": [
                    {
                        "description": "Model object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_llm-manager_internal_database_models.Model"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created model",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_llm-manager_internal_database_models.Model"
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/import": {
            "post": {
                "description": "Imports an LLM model definition from a YAML file or YAML content. Accepts multipart/form-data (file upload) or JSON body with a yaml_content field.",
                "consumes": [
                    "application/json",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models-utilities"
                ],
                "summary": "Import a model from YAML",
                "parameters": [
                    {
                        "description": "YAML content for import",
                        "name": "yaml_content",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.importRequest"
                        }
                    },
                    {
                        "type": "file",
                        "description": "YAML file to upload",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Imported model",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_llm-manager_internal_database_models.Model"
                        }
                    },
                    "400": {
                        "description": "Malformed request or invalid file",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/{slug}": {
            "get": {
                "description": "Returns a single LLM model identified by its slug.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Get a model by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model found",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_llm-manager_internal_database_models.Model"
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Performs a partial update of an existing model identified by its slug. Accepts a JSON map of field names to new values.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Update a model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial update fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated model",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_llm-manager_internal_database_models.Model"
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body or missing slug",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an LLM model identified by its slug. This is irreversible.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Delete a model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content\" \"Model deleted"
                    },
                    "400": {
                        "description": "Missing slug",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/{slug}/cache": {
            "delete": {
                "description": "Removes the HuggingFace model weights cache directory for the specified model. This frees disk space but requires re-downloading on next use.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models-utilities"
                ],
                "summary": "Clear HuggingFace model cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content\"\t\"Cache cleared"
                    },
                    "400": {
                        "description": "Missing slug or model has no HF repo",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to remove cache directory",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/{slug}/compose": {
            "get": {
                "description": "Generates a Docker Compose YAML configuration for the specified LLM model, including container settings, resource allocation, and vLLM engine configuration.",
                "produces": [
                    "text/x-yaml"
                ],
                "tags": [
                    "models-utilities"
                ],
                "summary": "Generate docker-compose for a model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated docker-compose YAML",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Missing slug",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Compose generation failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/{slug}/export": {
            "get": {
                "description": "Exports an LLM model from the database as a YAML document suitable for re-import to another instance.",
                "produces": [
                    "text/x-yaml"
                ],
                "tags": [
                    "models-utilities"
                ],
                "summary": "Export a model to YAML",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "YAML export document",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Missing slug",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Export or serialization failed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models/{slug}/info": {
            "get": {
                "description": "Returns an LLM model with its JSON string fields (LiteLLMParams, ModelInfo, Capabilities) parsed into typed objects.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "models"
                ],
                "summary": "Enriched model details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enriched model details",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ModelInfoResponse"
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rag": {
            "get": {
                "description": "Lists all RAG models (embedding and reranker) with their container status. When OData query parameters are present, returns a wrapped OData response with pagination metadata.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "List RAG models",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter expression (e.g. 'sub_type eq embedding')",
                        "name": "$filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Free-text search on name and slug",
                        "name": "$search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (e.g. 'name asc')",
                        "name": "$sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "$page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "$limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of fields to include",
                        "name": "$fields",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OData-wrapped response when OData params present",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ODataListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rag/start": {
            "post": {
                "description": "Starts the embedding and/or reranker container(s) for RAG pipelines. If slugs are omitted, starts the first available model of each type.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rag"
                ],
                "summary": "Start RAG containers",
                "parameters": [
                    {
                        "description": "Request with embed_slug and/or rerank_slug",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.StartRAGRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of started model slugs",
                        "schema": {
                            "$ref": "#/definitions/internal_api.StartRAGResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "No models available or model not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to start container",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.ODataListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/internal_api.ODataMeta"
                }
            }
        },
        "internal_api.ODataMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api.StartRAGRequest": {
            "type": "object",
            "properties": {
                "embed_slug": {
                    "type": "string"
                },
                "rerank_slug": {
                    "type": "string"
                }
            }
        },
        "internal_api.StartRAGResponse": {
            "type": "object",
            "properties": {
                "started": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "internal_api.importRequest": {
            "type": "object",
            "properties": {
                "yaml_content": {
                    "type": "string"
                }
            }
        }
    }
}